feat: Assert that only review issuer can delete it

+++ src/main/java/com/vivemedellin/valoracion_comentarios/review/application/commands/delete_review/DeleteReviewHandler.java
@@ -2,11 +2,12 @@ package com.vivemedellin.valoracion_comentarios.review.application.commands.dele
 
 import com.vivemedellin.valoracion_comentarios.review.exceptions.NotFoundReviewException;
 import com.vivemedellin.valoracion_comentarios.review.repository.ReviewRepository;
+import com.vivemedellin.valoracion_comentarios.shared.exceptions.ForbiddenAccessException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class DeleteReviewHandler {
+public class DeleteReviewHandler{
 
     private final ReviewRepository reviewRepository;
 
@@ -19,6 +20,7 @@ public class DeleteReviewHandler {
         var review = this.reviewRepository
                 .findById(command.getEventId())
                 .orElseThrow(() -> new NotFoundReviewException(command.getEventId()));
+        if(!review.getUserId().equals(command.getUserId())) throw new ForbiddenAccessException();
         this.reviewRepository.delete(review);
         return "Review deleted";
     }
